ubuntu 18.04 (#2956)
* fixes (warning removal mainly) for Ubuntu 18.04

* Prevent inlining of IdletimeTable::write_merge_regs

- gcc7 somehow inlines this into compilation units that don't have
  access to the template it calls (link time optimization?  polymorphic
  specialization?) which then causes a link error.

* add skeleton for testing compiling the compiler on different Ubuntu variants.

Adds a set of Docker files that install bf-p4c dependencies on a
barebones Ubuntu variant (16.04 and 18.04 for now. 14.04 is a
mess!). With these, we can deploy jobs that simply verify that the
compiler compiles. No testing necessary at this time.

TODO: resolve how we install bf-utils. Maybe another container on top
of the base image?

* update test failures due to new p4c tests

+++ scripts/pkg-src/ubuntu/mk_grpc_pkg.sh
@@ -2,6 +2,11 @@
 
 set -e
 
+pkg_install=false
+if [ $1 == '--install' ]; then
+    pkg_install=true
+fi
+
 mk_grpc_from_source() {
     git clone https://github.com/google/grpc && \
         cd grpc && \
@@ -120,9 +125,13 @@ builddir=$(mktemp --directory -t grpc_XXXXXX)
 mydir=$(realpath `dirname $0`)
 pushd $builddir
 mk_grpc_from_source
-set -x
-cp -r grpc-dev*.deb ${mydir}/${arch}
-popd
-rm -rf $builddir
+if $pkg_install == true; then
+    dpkg -i grpc-dev*.deb
+else
+    set -x
+    cp -r grpc-dev*.deb ${mydir}/${arch}
+    popd
+    rm -rf $builddir
 
-# dpkg-sig --sign builder ${mydir}/${arch}/grpc-dev*.deb
+    # dpkg-sig --sign builder ${mydir}/${arch}/grpc-dev*.deb
+fi
